requirements.txt
# Core dependencies for code-to-documentation system

# Database
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
alembic>=1.12.0

# LangGraph and LangChain
langgraph>=0.0.40
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.1.0

# Code Analysis
tree-sitter>=0.20.0
tree-sitter-python>=0.20.0
tree-sitter-javascript>=0.20.0
gitpython>=3.1.0

# Machine Learning and Embeddings
openai>=1.0.0
numpy>=1.24.0
sentence-transformers>=2.2.0
tiktoken>=0.5.0

# Text Processing
nltk>=3.8
spacy>=3.7.0

# API and Web Framework
fastapi>=0.100.0
uvicorn>=0.24.0
streamlit>=1.28.0

# Data Models and Validation
pydantic>=2.0.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
rich>=13.0.0
click>=8.1.0

# Development and Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0